fm - move Message, Transaction and Operation classes to their own package model to enable communications with Client App.
