+++ replicated-commit/src/test/java/edu/ucsb/rc/MultiDatacenterTest.java
@@ -12,7 +12,7 @@ import org.powermock.api.easymock.PowerMock;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import edu.ucsb.rc.network.Message;
+import edu.ucsb.rc.model.Message;
 import edu.ucsb.rc.network.NetworkHandler;
 import edu.ucsb.rc.protocols.PaxosAcceptsManager;
 import edu.ucsb.rc.protocols.TwoPhaseCommitManager;
