+++ replicated-commit/src/test/java/edu/ucsb/rc/network/ClientNetworkWorkerTest.java
@@ -11,7 +11,8 @@ import org.powermock.api.easymock.PowerMock;
 
 import edu.ucsb.rc.MultiDatacenter;
 import edu.ucsb.rc.Shard;
-import edu.ucsb.rc.transactions.Transaction;
+import edu.ucsb.rc.model.Message;
+import edu.ucsb.rc.model.Transaction;
 
 public class ClientNetworkWorkerTest {
